refactor(config): use getMemorySize() for size limit configs
Replace getInt() with getMemorySize().toBytes() for maxMessageSize,
httpMaxMessageSize and jsonRpcMaxMessageSize config parsing. This
enables human-readable size values (e.g. 4m, 128MB) while maintaining
backward compatibility with raw byte values.

- maxMessageSize (gRPC): keep int field, validate <= Integer.MAX_VALUE
  for gRPC API compatibility, add positive value validation
- httpMaxMessageSize / jsonRpcMaxMessageSize: widen to long, matching
  Jetty SizeLimitHandler's long parameter type
- Add config.conf documentation for all three size parameters

+++ common/src/main/java/org/tron/common/parameter/CommonParameter.java
@@ -216,10 +216,10 @@ public class CommonParameter {
   public int maxMessageSize;
   @Getter
   @Setter
-  public int httpMaxMessageSize;
+  public long httpMaxMessageSize;
   @Getter
   @Setter
-  public int jsonRpcMaxMessageSize;
+  public long jsonRpcMaxMessageSize;
   @Getter
   @Setter
   public int maxHeaderListSize;
